web/requirements.txt
Django
redis==2.10.3
django-redis==4.3.0
gunicorn==19.3.0
psycopg2==2.6

django-markdown-app
django-markdownx
pyembed
pyembed-markdown